% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lubridate.R
\name{add_weeks}
\alias{add_weeks}
\title{Append week timestamps fields in a data.frame}
\usage{
add_weeks(x, fields, ...)
}
\arguments{
\item{x}{The \link[base:data.frame]{base::data.frame} to be weekified.}

\item{fields}{(character vector) Columns to be weekified.}

\item{\dots}{Not used.}
}
\value{
A \link[base:data.frame]{base::data.frame} of the same dimensions as \code{code}.
}
\description{
Append week timestamps fields in a data.frame
}
