% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymize.R
\name{anonymize}
\alias{anonymize}
\title{Anonymize fields in a data.frame}
\usage{
anonymize(x, fields, ...)
}
\arguments{
\item{x}{The \link[base:data.frame]{base::data.frame} to be anonymized.}

\item{fields}{(character vector) Columns to be anonymized.}

\item{\dots}{Not used.}
}
\value{
A \link[base:data.frame]{base::data.frame} of the same dimensions as \code{code}.
}
\description{
Anonymize fields in a data.frame
}
