% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge_accounting.R
\name{as_sge_accounting}
\alias{as_sge_accounting}
\title{Coerce to an 'sge_accounting' Object}
\usage{
as_sge_accounting(x, ...)
}
\arguments{
\item{x}{The object to be coerced.}

\item{\ldots}{(optional) Addition arguments passed to the S3 methods.}
}
\description{
Coerce to an 'sge_accounting' Object
}
