% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available.R
\name{available.queues}
\alias{available.queues}
\title{Gets Queues that are Available For Jobs}
\usage{
\method{available}{queues}(df, ...)
}
\arguments{
\item{df}{A \code{queues} data frame.}

\item{\ldots}{Not used.}
}
\value{
Return the subset of queues that are on available hosts with a
known load and where either the 'short.q' or the 'long.q' queue
is available for jobs (= not overloaded).
}
\description{
Gets Queues that are Available For Jobs
}
