% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpu_info.R
\name{cpu_info}
\alias{cpu_info}
\title{Lists the CPU Information for the Current Machine}
\usage{
cpu_info()
}
\value{
A data.frame (tibble) with the two columns `field` and `value`.
}
\description{
Lists the CPU Information for the Current Machine
}
\details{
This function reads/queries `/proc/cpuinfo` on the current machine.
}
\examples{
info <- cpu_info()
print(info)

}
