% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disk_free.R
\name{disk_free}
\alias{disk_free}
\title{Report on File System Disk Usage}
\usage{
disk_free(mounts = NULL, local = FALSE)
}
\arguments{
\item{mounts}{A character vector of disks (mount points) to query.
If \code{NULL}, all disks are queried.}

\item{local}{If \code{TRUE}, only local disks are queried.}
}
\value{
A data.frame (tibble) where each row corresponds to a unique disk.
}
\description{
Report on File System Disk Usage
}
\details{
This function queries \code{df -P} (POSIX output format) on the system.
}
\examples{
## All disks
df <- disk_free()
print(df)

## All local disks
df <- disk_free(local = TRUE)
print(df)

## /tmp 
df <- disk_free(c("/tmp"))
print(df)

}
