% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_index_rows.R
\name{file_index_rows}
\alias{file_index_rows}
\title{Index the beginning of rows in a file}
\usage{
file_index_rows(pathname, newline = "\\n", bfr_size = 10000L)
}
\arguments{
\item{pathname}{(character) The file to be indexed.}

\item{newline}{(character) The character to scan for.}

\item{bfr_size}{(numeric) The number of bytes to read in each iteration.}
}
\value{
A numeric vector of file byte offsets that corresponds to the
beginning of a line, i.e. a position in the file that was preceeded
by a \code{newline} character.  The first line is at file byte offset
\code{0}, which is also always the first element in the returned vector.
}
\description{
Index the beginning of rows in a file
}
