% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qhost.R
\name{has_qhost}
\alias{has_qhost}
\title{Check Whether 'qhost' is Available on the System or not}
\usage{
has_qhost()
}
\value{
(logical)
\code{has_qhost()} returns TRUE if the \code{qhost} could be found, otherwise FALSE.
}
\description{
Check Whether 'qhost' is Available on the System or not
}
