% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host_info.R
\name{host_info}
\alias{host_info}
\title{Compile Host Information}
\usage{
host_info(system_info = NULL, qhost = NULL)
}
\arguments{
\item{system_info}{(optional) A named list as returned by [system_info()].}

\item{qhost}{(optional) A data.frame as returned by [qhost()].}
}
\value{
A data.frame (tibble) with information combined based
on `system_info` and `qhost`.
}
\description{
Compile Host Information
}
