% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host_table.R
\name{host_table}
\alias{host_table}
\title{Produce a Table of Static Host Information}
\usage{
host_table(host_info, format = TRUE)
}
\arguments{
\item{host_info}{A data.frame as returned by one or more
[host_info()] calls.}

\item{format}{If `TRUE`, the output is formatted to strings
with units and friendly column names.}
}
\value{
A data.frame.
}
\description{
Produce a Table of Static Host Information
}
