% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_jobs.R
\name{job_summary_by_status}
\alias{job_summary_by_status}
\title{Summarize Jobs by Status and Process Priority}
\usage{
job_summary_by_status(jobs, ...)
}
\arguments{
\item{jobs}{An 'sge_accounting' data.frame.}

\item{\ldots}{Currently not used.}
}
\value{
A named list with two matrices \code{total} and \code{fraction}.
}
\description{
Summarize Jobs by Status and Process Priority
}
