% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lscpu.R
\name{lscpu}
\alias{lscpu}
\title{Lists the CPU Information for the Current Machine}
\usage{
lscpu()
}
\value{
A data.frame (tibble) with the two columns `field` and `value`.
For most CPUs, there will be duplicated `field`:s.
}
\description{
Lists the CPU Information for the Current Machine
}
\details{
This function queries `lscpu` on the current machine.
}
\examples{
info <- lscpu()
print(info)

}
\seealso{
[cpu_info()]
}
