% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/on_hostname.R
\name{on_hostname}
\alias{on_hostname}
\title{Evaluates an R Expression on One or More Hosts}
\usage{
on_hostname(hostnames, expr, ...)
}
\arguments{
\item{hostnames}{A character vector of hostnames.}

\item{expr}{The \R expression to be evaluated on each host.}

\item{\ldots}{(optional) Additional arguments passed to [future::future()].}
}
\value{
A named list with names `hostnames`.
}
\description{
Evaluates an R Expression on One or More Hosts
}
\details{
This function utilizes the \pkg{future} framework and specifically
the \pkg{future.batchtools} package to evaluate the expression on
a particular host by submitting it via the SGE scheduler.
}
\examples{
\donttest{\dontrun{
q <- queues(filter = "available")
hostnames <- sort(unique(q$hostname))
si <- on_hostname(hostnames, system_info())
print(si)
}}

}
