% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_index.R
\name{open_file_at}
\alias{open_file_at}
\title{Open a file connection at given byte position}
\usage{
open_file_at(file, position = 0, open = "rb", auto_close = TRUE)
}
\arguments{
\item{file}{A pathname to a \file{*.index} file to be
created or read from.}

\item{position}{The file byte position where to position the connection
after being opened.}

\item{open}{(character) Mode to open the file connection in, cf.
\code{\link[base:connections]{base::file()}}.}

\item{auto_close}{(logical) If TRUE, then the connection is automatically
closed when garbage collected.}
}
\value{
A file \link[base:connections]{base::connection} moved to position \code{position}.
}
\description{
Open a file connection at given byte position
}
