% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge_accounting.R
\name{parse_category}
\alias{parse_category}
\title{Parse SGE Accounting 'category' Field}
\usage{
parse_category(x, ...)
}
\arguments{
\item{x}{An \code{sge_accounting} object.}

\item{\ldots}{(optional) Not used.}
}
\value{
A \code{tibble} data frame with columns corresponding to the requested
properties.
}
\description{
Parse SGE Accounting 'category' Field
}
