% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qhost.R
\name{qhost}
\alias{qhost}
\title{List Available Host and Their Current Statuses}
\usage{
qhost()
}
\value{
A data.frame (tibble) where each row corresponds to a unique host
(compute node).
}
\description{
List Available Host and Their Current Statuses
}
\details{
This function queries SGE's `qstat` on the system.
}
\examples{
\donttest{
qh <- qhost()
print(qh)

## Ignore compute nodes that are without load
qh <- subset(qh, !is.na(load))
print(qh)
}

}
