% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queues.R
\name{queues}
\alias{queues}
\title{List Available Queues and Their Properties}
\usage{
queues(filter = c("none", "available"), parse = TRUE)
}
\arguments{
\item{filter}{A predefined filter to apply, unless `"none"`.
If `"available"`, only queues on available hosts with a known load and
where either the 'short.q' or the 'long.q' queue is non-overloaded
are returned.}

\item{parse}{If `TRUE`, `qhost -f` columns that contain multiple properties
(e.g. 'states') are further parsed for easy access.}
}
\value{
A data.frame (tibble) where each row corresponds to a
unique queue on a particular host.
}
\description{
List Available Queues and Their Properties
}
\details{
This function queries SGE's `qstat -f` on the system.
}
\examples{
\donttest{
q <- queues()
print(q)

## Ignore queues whose nodes are disabled, without load, or those
## on developer and test nodes
q <- queues(filter = "available")
print(q)
}

}
