% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qstat.R
\name{read_qstat}
\alias{read_qstat}
\title{Get information on jobs currently on the SGE queue}
\usage{
read_qstat()
}
\value{
A \link[tibble:tibble]{tibble::tibble}
}
\description{
Get information on jobs currently on the SGE queue
}
