% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge_accounting.R
\name{read_raw_sge_accounting}
\alias{read_raw_sge_accounting}
\title{Read a tab-delimited SGE accounting file (without parsing it)}
\usage{
read_raw_sge_accounting(file, skip = 4L, ...)
}
\description{
Read a tab-delimited SGE accounting file (without parsing it)
}
\details{
The SGE \file{accounting} file is typically located in a subfolder of
the folder \file{$SGE_ROOT/$SGE_CELL/}.  On Wynton HPC, the pathname
is given by \code{sge_accounting_file()}.
}
