% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_ram.R
\name{round_ram}
\alias{round_ram}
\title{Round RAM Sizes to Nearest Possible RAM Size}
\usage{
round_ram(x, map = getOption("wyntonquery.round_ram.map", c(16, 32, 48, 64,
  96, 128, 256, 384, 512, 768, 1024, 2048)))
}
\arguments{
\item{x}{A numeric vector of RAM sizes to round.}

\item{map}{A numeric vector of RAM sizes to round.}
}
\value{
A numeric vector of length `length(x)` with
values in the `map` set.
}
\description{
Round RAM Sizes to Nearest Possible RAM Size
}
\examples{
x <- c(14.5, 506.2, 46.2)
print(x)
x2 <- round_ram(x)
print(x2)

}
