% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge_accounting.R
\name{sge_accounting_file}
\alias{sge_accounting_file}
\title{Locate the SGE Accounting File on the Current System}
\usage{
sge_accounting_file(
  filename = "accounting",
  path = do.call(file.path, args = as.list(c(Sys.getenv(c("SGE_ROOT", "SGE_CELL")),
    "common")))
)
}
\arguments{
\item{filename}{(character string) The name of the accounting file.}

\item{path}{(character string) The path to the accounting file.}
}
\value{
(character string) The pathname to the SGE accounting file.
If not found, and error is thrown.
}
\description{
Locate the SGE Accounting File on the Current System
}
