% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge_accounting.R
\name{sge_failed_codes}
\alias{sge_failed_codes}
\title{Table of SGE failed codes with descriptions}
\usage{
sge_failed_codes()
}
\value{
A tibble
}
\description{
Table of SGE failed codes with descriptions
}
