% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge_find_weeks.R
\name{sge_find_weeks}
\alias{sge_find_weeks}
\title{Index the SGE accounting file for ISO-8601 weeks}
\usage{
sge_find_weeks(
  file,
  index,
  until = NULL,
  n_max = Inf,
  delta = 1e+05,
  debug = FALSE
)
}
\arguments{
\item{file}{(character) The SGE \file{accounting} file to read.}

\item{index}{(numeric vector) File byte offsets of job entries.}

\item{until}{(character vector) If a week read that is part of this
set, the terminate the scanning.  The weeks should be specified
using \verb{\%GW\%V} week format, e.g. \verb{2019W01}.}

\item{n_max}{(numeric) The maximum number of weeks to read.}

\item{delta}{(numeric) Maximum number of rows to skip forward and
backword in the divide-and-conquer search strategy.}

\item{debug}{(logical) If TRUE, display debug messages.}
}
\value{
A named list of file-byte offsets of the first job with
and end time in the corresponding week.  The entries are named
using \verb{\%GW\%V} week format, e.g. \verb{2021W48}.
}
\description{
Index the SGE accounting file for ISO-8601 weeks
}
