% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge_make_week_index.R
\name{sge_make_week_index}
\alias{sge_make_week_index}
\title{Index the SGE accounting file for ISO-8601 weeks}
\usage{
sge_make_week_index(
  file,
  index,
  until = NULL,
  n_max = Inf,
  delta = 1e+05,
  debug = FALSE
)
}
\arguments{
\item{file}{(character) The SGE \file{accounting} file to read.}

\item{index}{(numeric vector) File byte offsets of job entries.}

\item{until}{(character vector) If a week read that is part of this
set, the terminate the scanning.  The weeks should be specified
using \verb{\%GW\%V} week format, e.g. \verb{2019W01}.}

\item{n_max}{(numeric) The maximum number of weeks to read.}

\item{delta}{(numeric) Maximum number of rows to skip forward and
backword in the divide-and-conquer search strategy.}

\item{debug}{(logical) If TRUE, display debug messages.}
}
\value{
A \link[tibble:tibble]{tibble::tibble} with fields:
\itemize{
\item \code{week} (character) - the ISO-8601 week name, e.g. \code{"2021W48"}
\item \code{nbr_of_jobs} (numeric) - number of jobs
\item \code{file_offset} (numeric) - the file byte offset
The week names uses \verb{\%GW\%V} week format.
}
}
\description{
Index the SGE accounting file for ISO-8601 weeks
}
