% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_info.R
\name{system_info}
\alias{system_info}
\title{Lists System Information for the Current Machine}
\usage{
system_info()
}
\value{
A named list with elements `hostname` (a character string),
`os_info`, `cpu_info`, and `disk_free` (data frames/tibbles).
}
\description{
Lists System Information for the Current Machine
}
\details{
This function reads/queries [base::Sys.info()], [disk_free()],
and [cpu_info()] on the current machine.
}
\examples{
info <- system_info()
print(info)

}
