% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beegfs_quota.R
\name{beegfs_quota}
\alias{beegfs_quota}
\title{Query BeeGFS for Disk Quota}
\usage{
beegfs_quota(which = "<current>", storagepool, what = c("user", "group"))
}
\arguments{
\item{which}{Subset of users or groups to be queried.
If NULL, quotas for all are queried.
If \verb{<current>}, then the current user or group is used.
If}

\item{storagepool}{A storage-pool ID (integer) to be queried.}

\item{what}{Either \code{"user"} or \code{"group"}}
}
\value{
A data.frame (tibble)
}
\description{
Query BeeGFS for Disk Quota
}
\details{
This function queries \code{beegfs-ctl} on the system.
}
\examples{
\donttest{\dontrun{
## The current user
quota <- beegfs_quota(what = "user", storagepool = 11L)
print(quota)

## The current group
quota <- beegfs_quota(what = "group", storagepool = 11L)
print(quota)

## All groups
quota <- beegfs_quota(which = NULL, what = "group", storagepool = 12L)
print(quota)
}}

}
