% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpu_info.R
\name{cpu_info}
\alias{cpu_info}
\title{Lists the CPU Information for the Current Machine}
\usage{
cpu_info()
}
\value{
A data.frame (tibble) with the two columns \code{field} and \code{value}.
For most CPUs, there will be duplicated \code{field}:s.
}
\description{
Lists the CPU Information for the Current Machine
}
\details{
This function reads/queries \code{/proc/cpuinfo} on the current machine.
}
\examples{
info <- cpu_info()
print(info)

}
\seealso{
\code{\link[=lscpu]{lscpu()}}
}
