% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host_info.R
\name{host_info}
\alias{host_info}
\title{Compile Host Information}
\usage{
host_info(system_info = NULL, qhost = TRUE)
}
\arguments{
\item{system_info}{(optional) A named list as returned by \code{\link[=system_info]{system_info()}}.}

\item{qhost}{(optional) A data.frame as returned by \code{\link[=qhost]{qhost()}}.
If \code{TRUE} (default), then \code{\link[=qhost]{qhost()}} is queried.
If \code{FALSE}, then queue information is not reported.}
}
\value{
A data.frame (tibble) with information combined based
on \code{system_info} and \code{qhost}.
}
\description{
Compile Host Information
}
