% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_jobs.R
\name{job_summary_by_status}
\alias{job_summary_by_status}
\title{Summarize Jobs by Status and Process Priority}
\usage{
job_summary_by_status(jobs, ...)
}
\arguments{
\item{jobs}{An 'sge_accounting' data.frame.}

\item{\ldots}{Currently not used.}
}
\value{
A named list with two matrices \code{total} and \code{fraction}.
}
\description{
Summarize Jobs by Status and Process Priority
}
\examples{
\donttest{\dontrun{
## Get all jobs that started during the last 30 days
jobs <- read_sge_accounting("accounting.csv", skip = 4L)
jobs <- subset(jobs, start_time >= Sys.time() - 30*24*3600)

# Summarize by status and process priority
stats <- job_summary_by_status(jobs)
print(stats)
# $total
#                      all         0         10         19
# success       1546740016 362692738 3389595.99 1180657682
# fail           104934943  80500076 1541827.09   22893040
# exceeded_h_rt   66864152  61657974    1116.87    5205061
# 
# $fraction
#                      all         0           10          19
# success       0.93646756 0.8183633 0.6873464179 0.980978749
# fail          0.06353244 0.1816367 0.3126535821 0.019021251
# exceeded_h_rt 0.04048263 0.1391222 0.0002264803 0.004324755
}}

}
