% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/on_hostname.R
\name{on_hostname}
\alias{on_hostname}
\title{Evaluates an R Expression on One or More Hosts}
\usage{
on_hostname(hostnames, expr, ..., on_error = c("stop", "asis"),
  cache = FALSE)
}
\arguments{
\item{hostnames}{A character vector of hostnames.}

\item{expr}{The \R expression to be evaluated on each host.}

\item{on_error}{A character string specifying how to handle errors.
If \code{"stop"} (default), then the error is relayed.
If \code{"asis"}, then the error (the \link[base:condition]{base::condition} object) is returned.}

\item{cache}{If \code{TRUE}, cached results on file are considered.}

\item{\ldots}{(optional) Additional arguments passed to \code{\link[future:future]{future::future()}}.}
}
\value{
A named list with names \code{hostnames}.
}
\description{
Evaluates an R Expression on One or More Hosts
}
\details{
This function utilizes the \pkg{future} framework and specifically
the \pkg{future.batchtools} package to evaluate the expression on
a particular host by submitting it via the SGE scheduler.
}
\examples{
\donttest{\dontrun{
q <- available(queues())
hostnames <- sort(unique(q$hostname))
si <- on_hostname(hostnames, system_info())
print(si)
}}

}
