% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queues.R
\name{queues}
\alias{queues}
\title{List Available Queues and Their Properties}
\usage{
queues(parse = TRUE)
}
\arguments{
\item{parse}{If \code{TRUE}, \code{qhost -f} columns that contain multiple properties
(e.g. 'states') are further parsed for easy access.}
}
\value{
A data.frame (tibble) where each row corresponds to a
unique queue on a particular host.
}
\description{
List Available Queues and Their Properties
}
\details{
This function queries SGE's \code{qstat -f} on the system.
}
\examples{
\donttest{
q <- queues()
print(q)

## Ignore queues whose nodes are disabled, without load, or those
## on developer and test nodes
q <- available(q)
print(q)
}

}
