% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qstat.R
\name{read_qstat}
\alias{read_qstat}
\title{Get information on jobs currently on the SGE queue}
\usage{
read_qstat(file = NULL)
}
\arguments{
\item{file}{A file, a connection, or NULL.  If NULL (default), then
the output of system call \verb{qstat -xml -u '*'} is read.}
}
\value{
A \link[tibble:tibble]{tibble::tibble}
}
\description{
Get information on jobs currently on the SGE queue
}
\examples{
pathname <- system.file("exdata", "qstat.xml", package = "wyntonquery")

qstat <- read_qstat(pathname)
print(qstat)

## Anonymize (although actually already anonymized)
qstat_anon <- anonymize(qstat)
print(qstat_anon)
}
