% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge_accounting.R
\name{read_raw_sge_accounting}
\alias{read_raw_sge_accounting}
\alias{write_raw_sge_accounting}
\alias{anonymize_raw_sge_accounting}
\title{Read a tab-delimited SGE accounting file (without parsing it)}
\usage{
read_raw_sge_accounting(file, skip = 4L, ...)

write_raw_sge_accounting(x, file, ...)

anonymize_raw_sge_accounting(x, fields = c("group", "owner", "project"))
}
\arguments{
\item{file}{(character) The SGE \file{accounting} file to read.}

\item{skip}{(integer) Number of lines to skip before parsing file content.}

\item{x}{(raw_sge_accounting) An \code{tibble} data frame of class
\code{raw_sge_accounting}.}

\item{fields}{(character vector) Columns to be anonymized.}

\item{\ldots}{(optional) Addition arguments passed to \code{\link[readr:read_delim]{readr::read_delim()}}.}
}
\description{
Read a tab-delimited SGE accounting file (without parsing it)
}
\details{
The SGE \file{accounting} file is typically located in a subfolder of
the folder \file{$SGE_ROOT/$SGE_CELL/}.  On Wynton HPC, the pathname
is given by \code{sge_accounting_file()}.
}
