% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge_accounting.R
\name{read_raw_sge_accounting}
\alias{read_raw_sge_accounting}
\alias{write_raw_sge_accounting}
\title{Read a tab-delimited SGE accounting file (without parsing it)}
\usage{
read_raw_sge_accounting(
  file,
  offset = 0,
  n_max = Inf,
  skip = if (is.character(file) && offset == 0) 4L else 0L,
  ...
)

write_raw_sge_accounting(x, file, header = attr(x, "header"), ...)
}
\arguments{
\item{file}{(character) The SGE \file{accounting} file to read.}

\item{offset}{The file byte position from where to start reading.}

\item{n_max}{(numeric) The maximum number of rows to read.}

\item{skip}{(integer) Number of lines to skip before parsing file content.}

\item{x}{(raw_sge_accounting) An \code{tibble} data frame of class
\code{raw_sge_accounting}.}

\item{header}{(character vector) Zero of more header lines to be written
at the top of the file.}

\item{\ldots}{(optional) Addition arguments passed to \code{\link[readr:read_delim]{readr::read_delim()}}.}
}
\description{
Read a tab-delimited SGE accounting file (without parsing it)
}
\details{
The SGE \file{accounting} file is typically located in a subfolder of
the folder \file{$SGE_ROOT/$SGE_CELL/}.  On Wynton HPC, the pathname
is given by \code{sge_accounting_file()}.
}
\examples{
pathname <- system.file("exdata", "accounting", package = "wyntonquery")

jobs <- read_raw_sge_accounting(pathname)
print(jobs)

## Anonymize (although actually already anonymized)
jobs_anon <- anonymize(jobs)
print(jobs_anon)
}
