% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge_accounting.R
\name{read_sge_accounting}
\alias{read_sge_accounting}
\title{Read an SGE accounting}
\usage{
read_sge_accounting(file = sge_accounting_file(), ...)
}
\description{
Read an SGE accounting
}
\section{Benchmarking}{

The 4.8 GB \file{accounting} on Wynton HPC takes ~120s to read.
}

