% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_info.R
\name{system_info}
\alias{system_info}
\title{Lists System Information for the Current Machine}
\usage{
system_info()
}
\value{
A named list with elements \code{hostname} (a character string),
\code{os_info}, \code{cpu_info}, and \code{disk_free} (data frames/tibbles).
}
\description{
Lists System Information for the Current Machine
}
\details{
This function reads/queries \code{\link[base:Sys.info]{base::Sys.info()}}, \code{\link[=disk_free]{disk_free()}},
and \code{\link[=cpu_info]{cpu_info()}} on the current machine.
}
\examples{
info <- system_info()
print(info)

}
